fix regression; factor out simplification helpers
