+++ liquid-fixpoint.cabal
@@ -76,6 +76,7 @@ library
                     Language.Fixpoint.Solver.Rewrite
                     Language.Fixpoint.Solver.Prettify
                     Language.Fixpoint.Solver.Sanitize
+                    Language.Fixpoint.Solver.Simplify
                     Language.Fixpoint.Solver.Solution
                     Language.Fixpoint.Solver.Solve
                     Language.Fixpoint.Solver.Stats
