+++ src/Language/Fixpoint/Solver/PLE.hs
@@ -32,6 +32,7 @@ import           Language.Fixpoint.SortCheck
 import           Language.Fixpoint.Graph.Deps             (isTarget)
 import           Language.Fixpoint.Solver.Common          (askSMT, toSMT)
 import           Language.Fixpoint.Solver.Sanitize        (symbolEnv)
+import           Language.Fixpoint.Solver.Simplify
 import           Language.Fixpoint.Solver.Rewrite
 
 import Language.REST.AbstractOC as OC
@@ -1043,8 +1044,13 @@ instance Simplifiable Expr where
       tx e
         | Just e' <- M.lookup e (icSimpl ictx)
         = e'
+      
+      tx (PAtom rel e1 e2) = applyBooleanFolding rel e1 e2
       tx (EBin bop e1 e2) = applyConstantFolding bop e1 e2
       tx (ENeg e)         = applyConstantFolding Minus (ECon (I 0)) e
+      tx (EApp e1 e2)
+        | isSetPred e1    = applySetFolding e1 e2
+
       tx (EApp (EVar f) a)
         | Just (dc, c)  <- L.lookup f (knConsts γ)
         , (EVar dc', _) <- splitEApp a
@@ -1061,46 +1067,6 @@ instance Simplifiable Expr where
         , dc == dc'
         = es!!i
       tx e = e
-
-applyConstantFolding :: Bop -> Expr -> Expr -> Expr
-applyConstantFolding bop e1 e2 =
-  case (e1, e2) of
-    (ECon (R left), ECon (R right)) ->
-      Mb.fromMaybe e (cfR bop left right)
-    (ECon (R left), ECon (I right)) ->
-      Mb.fromMaybe e (cfR bop left (fromIntegral right))
-    (ECon (I left), ECon (R right)) ->
-      Mb.fromMaybe e (cfR bop (fromIntegral left) right)
-    (ECon (I left), ECon (I right)) ->
-      Mb.fromMaybe e (cfI bop left right)
-    _ -> e
-  where
-
-    e = EBin bop e1 e2
-
-    getOp :: Num a => Bop -> Maybe (a -> a -> a)
-    getOp Minus    = Just (-)
-    getOp Plus     = Just (+)
-    getOp Times    = Just (*)
-    getOp RTimes   = Just (*)
-    getOp _        = Nothing
-
-    cfR :: Bop -> Double -> Double -> Maybe Expr
-    cfR bop left right = fmap go (getOp' bop)
-      where
-        go f = ECon $ R $ f left right
-
-        getOp' Div      = Just (/)
-        getOp' RDiv     = Just (/)
-        getOp' op       = getOp op
-
-    cfI :: Bop -> Integer -> Integer -> Maybe Expr
-    cfI bop left right = fmap go (getOp' bop)
-      where
-        go f = ECon $ I $ f left right
-
-        getOp' Mod = Just mod
-        getOp' op  = getOp op
 
 
 -------------------------------------------------------------------------------
